requirements.txt
streamlit
torch
pillow
opencv-python
numpy
matplotlib